[fix] Moved back-end to DeviceBindFromForm.cs class
