+++ src/EasyEPlanner.csproj
@@ -141,6 +141,7 @@
     <Compile Include="Editor.cs" />
     <Compile Include="EProjectManager.cs" />
     <Compile Include="ExcelRepoter.cs" />
+    <Compile Include="DeviceBindFromForm.cs" />
     <Compile Include="IO.cs" />
     <Compile Include="LogFrm.cs">
       <SubType>Form</SubType>
